Log generator: Fix hinting during overlap
When a header and a message overlap, in some cases we have to decide
which field capture to keep, the one in the header or the one in the
message. It keeps the most valuable one.

When the header capture is kept, the hints for this field have to
be discarded too.